Rename: Generate -> Haskell.Generate
